+++ GUI/src/javafxgui/controller/TTTController.java
@@ -1,6 +1,6 @@
 package javafxgui.controller;
 
-import javafxgui.view.GUIView;
+import javafxgui.javafxcomponents.GUIView;
 import jttt.core.game.GameMaker;
 import jttt.core.game.Game;
 import jttt.core.game.GameType;
