+++ GUI/test/javafxgui/controller/TTTControllerTest.java
@@ -3,11 +3,11 @@ package javafxgui.controller;
 import javafx.embed.swing.JFXPanel;
 import javafx.scene.Scene;
 import javafx.scene.layout.StackPane;
-import javafxgui.GUIGameMaker;
-import javafxgui.GUIGameMakerFake;
-import javafxgui.GUIPlayerFactory;
-import javafxgui.GUIPlayerFactoryFake;
-import javafxgui.view.GUIViewSpy;
+import javafxgui.gamemaker.GUIGameMaker;
+import javafxgui.gamemaker.GUIGameMakerFake;
+import javafxgui.players.GUIPlayerFactory;
+import javafxgui.players.GUIPlayerFactoryFake;
+import javafxgui.javafxcomponents.GUIViewSpy;
 import jttt.core.board.Mark;
 import jttt.core.players.Player;
 import org.junit.Before;
