+++ GUI/test/javafxgui/event/EventHandlerTest.java
@@ -2,7 +2,7 @@ package javafxgui.event;
 
 import javafx.scene.Scene;
 import javafx.scene.layout.StackPane;
-import javafxgui.view.GUIViewSpy;
+import javafxgui.javafxcomponents.GUIViewSpy;
 import org.junit.Before;
 import org.junit.Test;
